+++ spec/operation_resolver_spec.rb
@@ -121,7 +121,7 @@ RSpec.describe OpenapiFirst::OperationResolver do
     describe 'params' do
       it 'has allowed query string parameters' do
         expected_params = {
-          'tags' => ['foo']
+          tags: ['foo']
         }
         expect(MyApi).to receive(:find_pets) do |params, _res|
           expect(params).to eq expected_params
@@ -137,7 +137,7 @@ RSpec.describe OpenapiFirst::OperationResolver do
         }
 
         expected_params = {
-          'id' => '1'
+          id: '1'
         }.merge(pet)
 
         expect(MyApi).to receive(:update_pet) do |params, _res|