+++ spec/router_spec.rb
@@ -56,7 +56,7 @@ RSpec.describe OpenapiFirst::Router do
         get '/pets/1'
 
         params = last_request.env[OpenapiFirst::PATH_PARAMS]
-        expect(params).to eq('petId' => '1')
+        expect(params).to eq(petId: '1')
       end
 
       it 'does not add path parameters if not defined for operation' do
@@ -64,7 +64,7 @@ RSpec.describe OpenapiFirst::Router do
         get 'pets'
 
         params = last_request.env[OpenapiFirst::PATH_PARAMS]
-        expect(params).to be_nil
+        expect(params).to be_empty
       end
     end